+++ crates/nu-cli/src/commands/math/min.rs
@@ -1,7 +1,7 @@
+use crate::commands::math::reducers::{reducer_for, Reduce};
 use crate::commands::math::utils::run_with_function;
 use crate::commands::WholeStreamCommand;
 use crate::prelude::*;
-use crate::utils::data_processing::{reducer_for, Reduce};
 use nu_errors::ShellError;
 use nu_protocol::{Signature, UntaggedValue, Value};
