+++ crates/nu-cli/src/utils.rs
@@ -1,5 +1,4 @@
 pub mod data;
-pub mod data_processing;
 pub mod test_bins;
 
 use crate::path::canonicalize;
