--- crates/nu_plugin_chart/build.rs
@@ -1,3 +1,0 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    nu_build::build()
-}
