--- crates/nu_plugin_chart/src/lib.rs
@@ -1,4 +1,0 @@
-mod chart;
-mod nu;
-
-pub use chart::Chart;
