--- crates/nu_plugin_chart/src/main.rs
@@ -1,6 +1,0 @@
-use nu_plugin::serve_plugin;
-use nu_plugin_chart::Chart;
-
-fn main() {
-    serve_plugin(&mut Chart::new());
-}
